SPSC队列的话就是不会复现这个问题

+++ src/main/java/com/jfireframework/jnet/common/util/MPSCFixArray.java
@@ -86,7 +86,7 @@ abstract class Core extends Pad3
     }
 
     protected final    Object[] buffer;
-    protected final    int      mask;
+    protected final    long      mask;
     protected final    int[]    availableBuffers;
     protected final    int      indexShift;
     protected          long     consumerLimit;
